testing the game, renaming files, adding __init__
